requirements.txt
pandas
fbprophet
mysql-connector-python